requirements.txt
django-registration==0.8